+++ bindingcollectionadapter/src/main/java/me/tatarka/bindingcollectionadapter/BindingRecyclerViewAdapter.java
@@ -1,7 +1,6 @@
 package me.tatarka.bindingcollectionadapter;
 
 import android.databinding.DataBindingUtil;
-import android.databinding.ObservableArrayList;
 import android.databinding.ObservableList;
 import android.databinding.ViewDataBinding;
 import android.support.annotation.LayoutRes;
@@ -12,7 +11,6 @@ import android.view.LayoutInflater;
 import android.view.ViewGroup;
 
 import java.lang.ref.WeakReference;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
@@ -27,14 +25,18 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
     @NonNull
     private final ItemViewSelector<T> selector;
     private final WeakReferenceOnListChangedCallback<T> callback = new WeakReferenceOnListChangedCallback<>(this);
-    // This is what the recyclerview sees. It will only be modified on the main thread.
-    private final List<T> boundItems = new ArrayList<>();
-    private ObservableList<T> items;
+    private List<T> items;
     private LayoutInflater inflater;
+    private ItemIds<T> itemIds;
+    // Number of attached recyclerviews, we don't have to listen to notifications if there are none.
+    private int recyclerViewRefCount = 0;
 
     /**
      * Constructs a new instance with the given {@link ItemView}.
+     *
+     * @deprecated Use {@link #BindingRecyclerViewAdapter(ItemViewArg)} instead.
      */
+    @Deprecated
     public BindingRecyclerViewAdapter(@NonNull ItemView itemView) {
         this.itemView = itemView;
         this.selector = BaseItemViewSelector.empty();
@@ -42,7 +44,10 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
 
     /**
      * Constructs a new instance with the given {@link ItemViewSelector}.
+     *
+     * @deprecated use {@link #BindingRecyclerViewAdapter(ItemViewArg)} intead.
      */
+    @Deprecated
     public BindingRecyclerViewAdapter(@NonNull ItemViewSelector<T> selector) {
         this.itemView = new ItemView();
         this.selector = selector;
@@ -58,39 +63,31 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
 
     @Override
     public void setItems(@Nullable Collection<T> items) {
-        if (this.items == items) {
-            return;
-        }
-
-        if (this.items != null) {
-            this.items.removeOnListChangedCallback(callback);
-            this.boundItems.clear();
-            notifyItemRangeRemoved(0, this.items.size());
-        }
-
-        if (items instanceof ObservableList) {
-            this.items = (ObservableList<T>) items;
-            this.boundItems.addAll(items);
-            notifyItemRangeInserted(0, this.items.size());
-            this.items.addOnListChangedCallback(callback);
-        } else if (items != null) {
-            this.items = new ObservableArrayList<>();
-            this.items.addOnListChangedCallback(callback);
-            this.items.addAll(items);
-        } else {
-            this.items = null;
-        }
+        setItems(Utils.getListFromCollection(items));
     }
 
     @Override
-    @Deprecated
-    public ObservableList<T> getItems() {
-        return items;
+    public void setItems(@Nullable List<T> items) {
+        if (this.items == items) {
+            return;
+        }
+        // If a recyclerview is listening, set up listeners. Otherwise wait until one is attached.
+        // No need to make a sound if nobody is listening right?
+        if (recyclerViewRefCount > 0) {
+            if (this.items instanceof ObservableList) {
+                ((ObservableList<T>) this.items).removeOnListChangedCallback(callback);
+            }
+            if (items instanceof ObservableList) {
+                ((ObservableList<T>) items).addOnListChangedCallback(callback);
+            }
+        }
+        this.items = items;
+        notifyDataSetChanged();
     }
 
     @Override
     public T getAdapterItem(int position) {
-        return boundItems.get(position);
+        return items.get(position);
     }
 
     @Override
@@ -103,17 +100,25 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
         if (bindingVariable != ItemView.BINDING_VARIABLE_NONE) {
             boolean result = binding.setVariable(bindingVariable, item);
             if (!result) {
-                BindingCollectionAdapters.throwMissingVariable(binding, bindingVariable, layoutRes);
+                Utils.throwMissingVariable(binding, bindingVariable, layoutRes);
             }
             binding.executePendingBindings();
         }
     }
 
     @Override
+    public void onAttachedToRecyclerView(RecyclerView recyclerView) {
+        if (recyclerViewRefCount == 0 && items != null && items instanceof ObservableList) {
+            ((ObservableList<T>) items).addOnListChangedCallback(callback);
+        }
+        recyclerViewRefCount += 1;
+    }
+
+    @Override
     public void onDetachedFromRecyclerView(RecyclerView recyclerView) {
-        callback.cancel();
-        if (items != null) {
-            items.removeOnListChangedCallback(callback);
+        recyclerViewRefCount -= 1;
+        if (recyclerViewRefCount == 0 && items != null && items instanceof ObservableList) {
+            ((ObservableList<T>) items).removeOnListChangedCallback(callback);
         }
     }
 
@@ -128,19 +133,33 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
 
     @Override
     public final void onBindViewHolder(ViewHolder viewHolder, int position) {
-        T item = boundItems.get(position);
+        T item = items.get(position);
         onBindBinding(viewHolder.binding, itemView.getBindingVariable(), itemView.getLayoutRes(), position, item);
     }
 
     @Override
     public int getItemViewType(int position) {
-        selector.select(itemView, position, boundItems.get(position));
+        selector.select(itemView, position, items.get(position));
         return itemView.getLayoutRes();
+    }
+
+    /**
+     * Set the item id's for the items. If not null, this will set {@link
+     * android.support.v7.widget.RecyclerView.Adapter#setHasStableIds(boolean)} to true.
+     */
+    public void setItemIds(@Nullable ItemIds<T> itemIds) {
+        this.itemIds = itemIds;
+        setHasStableIds(itemIds != null);
     }
 
     @Override
     public int getItemCount() {
-        return boundItems.size();
+        return items == null ? 0 : items.size();
+    }
+
+    @Override
+    public long getItemId(int position) {
+        return itemIds == null ? position : itemIds.getItemId(position, items.get(position));
     }
 
     public static class ViewHolder extends RecyclerView.ViewHolder {
@@ -152,15 +171,11 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
         }
     }
 
-    private static class WeakReferenceOnListChangedCallback<T> extends BaseOnListChangedCallback<T> {
+    private static class WeakReferenceOnListChangedCallback<T> extends ObservableList.OnListChangedCallback<ObservableList<T>> {
         final WeakReference<BindingRecyclerViewAdapter<T>> adapterRef;
 
         WeakReferenceOnListChangedCallback(BindingRecyclerViewAdapter<T> adapter) {
             this.adapterRef = new WeakReference<>(adapter);
-        }
-
-        public void cancel() {
-            handler.removeCallbacksAndMessages(null);
         }
 
         @Override
@@ -169,18 +184,8 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
             if (adapter == null) {
                 return;
             }
-            final List<T> changedItems = new ArrayList<>(adapter.items);
-            onMainThread(new OnMainThread() {
-                @Override
-                public void onMainThread() {
-                    BindingRecyclerViewAdapter<T> adapter = adapterRef.get();
-                    if (adapter != null) {
-                        adapter.boundItems.clear();
-                        adapter.boundItems.addAll(changedItems);
-                        adapter.notifyDataSetChanged();
-                    }
-                }
-            });
+            Utils.ensureChangeOnMainThread();
+            adapter.notifyDataSetChanged();
         }
 
         @Override
@@ -189,22 +194,8 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
             if (adapter == null) {
                 return;
             }
-            final List<T> changedItems = new ArrayList<>(itemCount);
-            for (int i = positionStart; i < positionStart + itemCount; i++) {
-                changedItems.add(adapter.items.get(positionStart));
-            }
-            onMainThread(new OnMainThread() {
-                @Override
-                public void onMainThread() {
-                    BindingRecyclerViewAdapter<T> adapter = adapterRef.get();
-                    if (adapter != null) {
-                        for (int i = positionStart; i < positionStart + itemCount; i++) {
-                            adapter.boundItems.set(i, changedItems.get(i - positionStart));
-                        }
-                        adapter.notifyItemRangeChanged(positionStart, itemCount);
-                    }
-                }
-            });
+            Utils.ensureChangeOnMainThread();
+            adapter.notifyItemRangeChanged(positionStart, itemCount);
         }
 
         @Override
@@ -213,22 +204,8 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
             if (adapter == null) {
                 return;
             }
-            final List<T> changedItems = new ArrayList<>(itemCount);
-            for (int i = positionStart; i < positionStart + itemCount; i++) {
-                changedItems.add(adapter.items.get(i));
-            }
-            onMainThread(new OnMainThread() {
-                @Override
-                public void onMainThread() {
-                    BindingRecyclerViewAdapter<T> adapter = adapterRef.get();
-                    if (adapter != null) {
-                        for (int i = positionStart; i < positionStart + itemCount; i++) {
-                            adapter.boundItems.add(i, changedItems.get(i - positionStart));
-                        }
-                        adapter.notifyItemRangeInserted(positionStart, itemCount);
-                    }
-                }
-            });
+            Utils.ensureChangeOnMainThread();
+            adapter.notifyItemRangeInserted(positionStart, itemCount);
         }
 
         @Override
@@ -237,25 +214,10 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
             if (adapter == null) {
                 return;
             }
-            final List<T> changedItems = new ArrayList<>(itemCount);
-            for (int i = fromPosition; i < fromPosition + itemCount; i++) {
-                changedItems.add(adapter.items.get(i));
+            Utils.ensureChangeOnMainThread();
+            for (int i = 0; i < itemCount; i++) {
+                adapter.notifyItemMoved(fromPosition + i, toPosition + i);
             }
-            onMainThread(new OnMainThread() {
-                @Override
-                public void onMainThread() {
-                    BindingRecyclerViewAdapter<T> adapter = adapterRef.get();
-                    if (adapter != null) {
-                        for (int i = fromPosition + itemCount - 1; i >= fromPosition; i--) {
-                            adapter.boundItems.remove(fromPosition);
-                        }
-                        adapter.boundItems.addAll(toPosition, changedItems);
-                        for (int i = 0; i < itemCount; i++) {
-                            adapter.notifyItemMoved(fromPosition + i, toPosition + i);
-                        }
-                    }
-                }
-            });
         }
 
         @Override
@@ -264,18 +226,12 @@ public class BindingRecyclerViewAdapter<T> extends RecyclerView.Adapter<BindingR
             if (adapter == null) {
                 return;
             }
-            onMainThread(new OnMainThread() {
-                @Override
-                public void onMainThread() {
-                    BindingRecyclerViewAdapter<T> adapter = adapterRef.get();
-                    if (adapter != null) {
-                        for (int i = positionStart + itemCount - 1; i >= positionStart; i--) {
-                            adapter.boundItems.remove(i);
-                        }
-                        adapter.notifyItemRangeRemoved(positionStart, itemCount);
-                    }
-                }
-            });
+            Utils.ensureChangeOnMainThread();
+            adapter.notifyItemRangeRemoved(positionStart, itemCount);
         }
+    }
+
+    public interface ItemIds<T> {
+        long getItemId(int position, T item);
     }
 }